prac/google-form-rpa-agent/requirements.txt
playwright==1.47.0
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=5.2.2
google-generativeai>=0.7.2
python-dotenv>=1.0.1
pydantic>=2.8.2
